cleanup accountbalance
